# Improve the MinGW builds and simplify bundling the static library dependencies.