--- a/.github/workflows/build-rtools40.yml
+++ b/.github/workflows/build-rtools40.yml
@@ -17,6 +17,13 @@ jobs:
       - uses: actions/checkout@v3
         with:
           fetch-depth: 1
+      # Configure required environment variables for vcpkg to use
+      # GitHub's Action Cache
+      - uses: actions/github-script@v6
+        with:
+          script: |
+            core.exportVariable('ACTIONS_CACHE_URL', process.env.ACTIONS_CACHE_URL || '');
+            core.exportVariable('ACTIONS_RUNTIME_TOKEN', process.env.ACTIONS_RUNTIME_TOKEN || '');
       - name: Install rtools40 if needed
         uses: r-windows/install-rtools@master
       - name: Building TileDB with Rtools40
@@ -27,6 +34,7 @@ jobs:
         env:
           TILEDB_HOME: ${{ github.workspace }}
           MINGW_ARCH: ${{ matrix.msystem }}
+          VCPKG_BINARY_SOURCES: 'clear;x-gha,readwrite'
         shell: c:\rtools40\usr\bin\bash.exe --login {0}
       - name: "Upload binaries"
         uses: actions/upload-artifact@v3