--- a/.github/workflows/mingw-w64-tiledb/PKGBUILD
+++ b/.github/workflows/mingw-w64-tiledb/PKGBUILD
@@ -26,9 +26,9 @@ build() {
   if [ "$CARCH" == "i686" ]; then
   export CFLAGS="-mfpmath=sse -msse2"
   export CXXFLAGS="-mfpmath=sse -msse2"
-  vcpkg_triplet="x86-mingw-static"
+  vcpkg_triplet="x86-mingw-static-release"
   else
-  vcpkg_triplet="x64-mingw-static"
+  vcpkg_triplet="x64-mingw-static-release"
   fi