--- a/.github/workflows/mingw-w64-tiledb/PKGBUILD
+++ b/.github/workflows/mingw-w64-tiledb/PKGBUILD
@@ -39,6 +39,7 @@ build() {
     -DBUILD_SHARED_LIBS=OFF \
     -DCMAKE_BUILD_TYPE=Release \
     -DTILEDB_STATIC=ON \
+    -DTILEDB_INSTALL_STATIC_DEPS=OFF \
     -DTILEDB_S3=ON \
     -DCOMPILER_SUPPORTS_AVX2=OFF \
     -DTILEDB_SKIP_S3AWSSDK_DIR_LENGTH_CHECK=ON \
@@ -60,6 +61,8 @@ build() {
 package() {
   cd ${source_dir}/build-${MINGW_CHOST}
   make DESTDIR="${pkgdir}" -C tiledb install
+  # Copy vcpkg_installed to the mingw package.
+  cp -r vcpkg_installed/${vcpkg_triplet}/. ${pkgdir}/${MINGW_ARCH}
 }
 
 check (){