--- a/.github/workflows/mingw-w64-tiledb/system-ports/aws-sdk-cpp/portfile.cmake
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/aws-sdk-cpp/portfile.cmake
@@ -1,1 +1,0 @@
-set(VCPKG_POLICY_EMPTY_PACKAGE enabled)