--- a/.github/workflows/mingw-w64-tiledb/system-ports/bzip2/vcpkg.json
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/bzip2/vcpkg.json
@@ -1,9 +1,0 @@
-{
-  "name": "bzip2",
-  "version-string": "system",
-  "features": {
-    "tool": {
-      "description": "Builds bzip2 executable"
-    }
-  }
-}