--- a/.github/workflows/mingw-w64-tiledb/system-ports/curl/portfile.cmake
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/curl/portfile.cmake
@@ -1,1 +1,0 @@
-set(VCPKG_POLICY_EMPTY_PACKAGE enabled)