--- a/.github/workflows/mingw-w64-tiledb/system-ports/lz4/portfile.cmake
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/lz4/portfile.cmake
@@ -1,1 +1,0 @@
-set(VCPKG_POLICY_EMPTY_PACKAGE enabled)