--- a/.github/workflows/mingw-w64-tiledb/system-ports/lz4/vcpkg.json
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/lz4/vcpkg.json
@@ -1,4 +1,0 @@
-{
-  "name": "lz4",
-  "version-string": "system"
-}