--- a/.github/workflows/mingw-w64-tiledb/system-ports/zlib/vcpkg.json
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/zlib/vcpkg.json
@@ -1,4 +1,0 @@
-{
-  "name": "zlib",
-  "version-string": "system"
-}