--- a/.github/workflows/mingw-w64-tiledb/system-ports/zstd/vcpkg.json
+++ b/.github/workflows/mingw-w64-tiledb/system-ports/zstd/vcpkg.json
@@ -1,4 +1,0 @@
-{
-  "name": "zstd",
-  "version-string": "system"
-}