… .github/workflows/mingw-w64-tiledb/system-ports/aws-sdk-cpp/vcpkg.json
… .github/workflows/mingw-w64-tiledb/system-ports/curl/vcpkg.json
… cmake/Options/BuildOptions.cmake
… cmake/TileDB-Superbuild.cmake
… ports/triplets/x64-mingw-static-release.cmake
… ports/triplets/x86-mingw-static-release.cmake