[SC-35901](https://app.shortcut.com/tiledb-inc/story/35901/copy-the-vcpkg-installed-directory-to-the-mingw-package)

As a preparation towards stopping bundling our static library dependencies (which would lead to significant simplifications in the build system), this PR updates the MinGW builds (which supply the binaries the R API uses on Windows) to stop relying on that automatic bundling. Our approach becomes more blunt, just copying the contents of the build directory's `vcpkg_installed` directory, where vcpkg places all libraries we might need.

Nothing (fingers crossed) should change on the side of the R API. These are the libraries that are bundled before and after this PR:

<details>
<summary>Before</summary>

```
-rw-r--r--  0 root   root   267222 Οκτ 19 16:08 ucrt64/lib/libaws-c-common.a
-rw-r--r--  0 root   root    28638 Οκτ 19 16:08 ucrt64/lib/libaws-c-event-stream.a
-rw-r--r--  0 root   root    47078 Οκτ 19 16:08 ucrt64/lib/libaws-checksums.a
-rw-r--r--  0 root   root  1540524 Οκτ 19 16:08 ucrt64/lib/libaws-cpp-sdk-cognito-identity.a
-rw-r--r--  0 root   root  2315096 Οκτ 19 16:08 ucrt64/lib/libaws-cpp-sdk-core.a
-rw-r--r--  0 root   root   319782 Οκτ 19 16:08 ucrt64/lib/libaws-cpp-sdk-identity-management.a
-rw-r--r--  0 root   root  5501212 Οκτ 19 16:08 ucrt64/lib/libaws-cpp-sdk-s3.a
-rw-r--r--  0 root   root   676178 Οκτ 19 16:08 ucrt64/lib/libaws-cpp-sdk-sts.a
-rw-r--r--  0 root   root    69652 Οκτ 19 16:08 ucrt64/lib/libbz2.a
-rw-r--r--  0 root   root   452820 Οκτ 19 16:08 ucrt64/lib/libfmt.a
-rw-r--r--  0 root   root   241756 Οκτ 19 16:08 ucrt64/lib/liblibmagic.a
-rw-r--r--  0 root   root   121288 Οκτ 19 16:08 ucrt64/lib/liblz4.a
-rw-r--r--  0 root   root   785406 Οκτ 19 16:08 ucrt64/lib/libpcre2-8.a
-rw-r--r--  0 root   root     3934 Οκτ 19 16:08 ucrt64/lib/libpcre2-posix.a
-rw-r--r--  0 root   root    33576 Οκτ 19 16:08 ucrt64/lib/libsharpyuv.a
-rw-r--r--  0 root   root  1404130 Οκτ 19 16:08 ucrt64/lib/libspdlog.a
-rw-r--r--  0 root   root   340224 Οκτ 19 16:08 ucrt64/lib/libtiledb.dll.a
-rw-r--r--  0 root   root 41120496 Οκτ 19 16:08 ucrt64/lib/libtiledbstatic.a
-rw-r--r--  0 root   root   881632 Οκτ 19 16:08 ucrt64/lib/libwebp.a
-rw-r--r--  0 root   root   485960 Οκτ 19 16:08 ucrt64/lib/libwebpdecoder.a
-rw-r--r--  0 root   root    14684 Οκτ 19 16:08 ucrt64/lib/libwebpdemux.a
-rw-r--r--  0 root   root    46538 Οκτ 19 16:08 ucrt64/lib/libwebpmux.a
-rw-r--r--  0 root   root   102320 Οκτ 19 16:08 ucrt64/lib/libz.a
-rw-r--r--  0 root   root   814174 Οκτ 19 16:08 ucrt64/lib/libzstd.a
```
</details>

<details>
<summary>After</summary>

```
-rw-r--r--  0 root   root   321018 Οκτ 20 14:33 ucrt64/lib/libaws-c-auth.a
-rw-r--r--  0 root   root    78040 Οκτ 20 14:33 ucrt64/lib/libaws-c-cal.a
-rw-r--r--  0 root   root   356172 Οκτ 20 14:33 ucrt64/lib/libaws-c-common.a
-rw-r--r--  0 root   root     8570 Οκτ 20 14:33 ucrt64/lib/libaws-c-compression.a
-rw-r--r--  0 root   root   121990 Οκτ 20 14:33 ucrt64/lib/libaws-c-event-stream.a
-rw-r--r--  0 root   root   539052 Οκτ 20 14:33 ucrt64/lib/libaws-c-http.a
-rw-r--r--  0 root   root   417204 Οκτ 20 14:33 ucrt64/lib/libaws-c-io.a
-rw-r--r--  0 root   root   415280 Οκτ 20 14:33 ucrt64/lib/libaws-c-mqtt.a
-rw-r--r--  0 root   root   242642 Οκτ 20 14:33 ucrt64/lib/libaws-c-s3.a
-rw-r--r--  0 root   root   130632 Οκτ 20 14:33 ucrt64/lib/libaws-c-sdkutils.a
-rw-r--r--  0 root   root    39962 Οκτ 20 14:33 ucrt64/lib/libaws-checksums.a
-rw-r--r--  0 root   root   448814 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-access-management.a
-rw-r--r--  0 root   root   849090 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-cognito-identity.a
-rw-r--r--  0 root   root  3630884 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-core.a
-rw-r--r--  0 root   root  4055126 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-iam.a
-rw-r--r--  0 root   root   346146 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-identity-management.a
-rw-r--r--  0 root   root  3799166 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-s3.a
-rw-r--r--  0 root   root   495284 Οκτ 20 14:33 ucrt64/lib/libaws-cpp-sdk-sts.a
-rw-r--r--  0 root   root  1237918 Οκτ 20 14:33 ucrt64/lib/libaws-crt-cpp.a
-rw-r--r--  0 root   root    98636 Οκτ 20 14:33 ucrt64/lib/libbz2.a
-rw-r--r--  0 root   root   452820 Οκτ 20 14:33 ucrt64/lib/libfmt.a
-rw-r--r--  0 root   root   241756 Οκτ 20 14:33 ucrt64/lib/liblibmagic.a
-rw-r--r--  0 root   root   212328 Οκτ 20 14:33 ucrt64/lib/liblz4.a
-rw-r--r--  0 root   root   304626 Οκτ 20 14:33 ucrt64/lib/liblzma.a
-rw-r--r--  0 root   root   719352 Οκτ 20 14:33 ucrt64/lib/libpcre2-16.a
-rw-r--r--  0 root   root   689066 Οκτ 20 14:33 ucrt64/lib/libpcre2-32.a
-rw-r--r--  0 root   root   785406 Οκτ 20 14:33 ucrt64/lib/libpcre2-8.a
-rw-r--r--  0 root   root     3934 Οκτ 20 14:33 ucrt64/lib/libpcre2-posix.a
-rw-r--r--  0 root   root    33576 Οκτ 20 14:33 ucrt64/lib/libsharpyuv.a
-rw-r--r--  0 root   root  1404130 Οκτ 20 14:33 ucrt64/lib/libspdlog.a
-rw-r--r--  0 root   root   340224 Οκτ 20 14:33 ucrt64/lib/libtiledb.dll.a
-rw-r--r--  0 root   root 41075680 Οκτ 20 14:33 ucrt64/lib/libtiledbstatic.a
-rw-r--r--  0 root   root   881632 Οκτ 20 14:33 ucrt64/lib/libwebp.a
-rw-r--r--  0 root   root   485960 Οκτ 20 14:33 ucrt64/lib/libwebpdecoder.a
-rw-r--r--  0 root   root    14684 Οκτ 20 14:33 ucrt64/lib/libwebpdemux.a
-rw-r--r--  0 root   root    46538 Οκτ 20 14:33 ucrt64/lib/libwebpmux.a
-rw-r--r--  0 root   root   123968 Οκτ 20 14:33 ucrt64/lib/libzlib.a
-rw-r--r--  0 root   root  1227472 Οκτ 20 14:33 ucrt64/lib/libzstd.a
```
</details>

Notice that no libraries have stoppped from being bundled. Because we now bundle libraries we don't actually need, the package's size increased:

| |Uncompressed|Compressed|
|-|------------|----------|
|Before|73.1MB|9.5MB|
|After|112.5MB|13.1MB|

Also the package now includes some useless files like around one thousand headers for the AWS SDK, but [they seem to be excluded from being uploaded](https://github.com/TileDB-Inc/rwinlib-tiledb/blob/aab71de1dd9d1c3b25ed2b1c6c6274c28e5e4f7f/update.sh#L33).

I also made some improvement that should speed-up the MinGW builds (build dependencies in Release mode only, cache them and disable tests).

---
TYPE: IMPROVEMENT
DESC: Improve building the Windows binaries for the R API.