@eddelbuettel 2023-10-27
Might be easier to just borrow a few minutes from a colleague using Windows and R, and I have a "volunteer" in mind.

Does the PR produce artifacts someone can access?