@eddelbuettel 2023-10-27
@mojaveazure was kind enough to give it a whirl but seems to have uncovered that this build was up to date with some recent merges in dev and some of the enumerations extensions we need in the current version of the R package are missing -- so a build attempt failed.  Another try with the released version of the R package should shed more light.