@eddelbuettel 2023-10-27
@teo-tsirpanis :  There is a lot of 'shrapnel' @mojaveazure is seeing in AWS linkage.  Example `undefined reference to `aws_mqtt_library_init'`.  We never needed an MQTT library (as that is AFAIK a messaging service).  Any idea if the set of libraries could be reduced?   Else we have to link with 'all of them' and given the possible link order permutations that could be tedious to sort out.