@teo-tsirpanis 2023-10-27
2.18 has updated the AWS SDK and the new version [unconditionally calls `aws_mqtt_library_init` when initializing the `aws-crt-cpp` library](https://github.com/awslabs/aws-crt-cpp/blob/f098dcbdbe6226f5b8e6f7424b4b38143dd3b820/source/Api.cpp#L44).

You will have to update the linked libraries in `Makevars.win`. Here's the list of imported AWS SDK targets in the installed CMake config, to help you with the ordering:

```
aws-cpp-sdk-identity-management
aws-cpp-sdk-cognito-identity
aws-cpp-sdk-sts
aws-cpp-sdk-s3
aws-cpp-sdk-core
AWS::aws-crt-cpp
AWS::aws-c-auth
AWS::aws-c-cal
AWS::aws-c-common
AWS::aws-c-compression
AWS::aws-c-event-stream
AWS::aws-c-http
AWS::aws-c-io
AWS::aws-c-mqtt
AWS::aws-c-s3
AWS::aws-checksums
AWS::aws-c-sdkutils
```