@eddelbuettel 2023-10-27
We will try to work it out. On Linux `cmake` hides that from me.